feat: commoConstants file renamed to stageConstants
