+++ frontend/src/community/crm/constants/stageConstants.ts
